services/identity_service/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.7.9
charset-normalizer==3.4.2
Django==5.2.3
django-debug-toolbar==5.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.7.0
geoip2==5.1.0
idna==3.10
maxminddb==2.7.0
multidict==6.6.3
propcache==0.3.2
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.4
setuptools==78.1.1
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
wheel==0.45.1
yarl==1.20.1
